papers/QCNN_data_classification/requirements.txt
# Photonic / quantum stack
merlinquantum

# Core scientific stack
scipy
scikit-learn
torchvision

# Utilities
matplotlib
pandas
pytest
